address gcc6 build error
With gcc6, compiling fails with `stdlib.h: No such file or directory`,
as including '-isystem /usr/include' breaks with gcc6, cf.,
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=70129.

This commit addresses this issue for this package in the same way
it was addressed in various other ROS packages. A list of related
commits and pull requests is at:

  https://github.com/ros/rosdistro/issues/12783

The SYSTEM attribute for the include directories was added in commit
188f8fe9 on 2012-11-10 when the packages were changed to use catkin as
build system. The exact reason for using the SYSTEM attribute cannot be
inferred from that commit.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@oss.bmw-carit.de>